>  compile 191220000

+++ nemu/src/monitor/debug/expr.c
@@ -219,8 +219,8 @@ word_t eval(Token *tokens, uint32_t left, uint32_t right, bool *success) {
 
     search_main_op(tokens, left, right, success, &op, &op_type);
 
-    uint64_t val1 = eval(tokens, left, op - 1, success);
-    uint64_t val2 = eval(tokens, op + 1, right, success);
+    uint32_t val1 = eval(tokens, left, op - 1, success);
+    uint32_t val2 = eval(tokens, op + 1, right, success);
 
     switch (op_type) {
       case '+': return val1 + val2;
